+++ src/module.cpp
@@ -4,7 +4,7 @@
 Napi::Object init(Napi::Env env, Napi::Object exports)
 {
     exports.Set(Napi::String::New(env, "composite"), Napi::Function::New(env, vtile::composite));
-    exports.Set(Napi::String::New(env, "internationalize"), Napi::Function::New(env, vtile::internationalize));
+    exports.Set(Napi::String::New(env, "localize"), Napi::Function::New(env, vtile::localize));
     return exports;
 }